--- src/components/dashboard/header.js
@@ -1,3 +1,0 @@
-export const Header = () => {
-  return <h3>Header</h3>;
-};
