+++ src/components/privateRoute.js
@@ -1,9 +1,8 @@
 import { useSelector } from 'react-redux';
-import { Outlet, Navigate } from 'react-router-dom';
+import { Navigate } from 'react-router-dom';
 
 export const PrivateRoute = ({ children }) => {
-  // const isLogged = useSelector(state => state.auth.isLogged);
-  const isLogged = true;
+  const isLogged = useSelector(state => state.auth.isLogged);
   console.log('Private Route');
   return <>{isLogged ? children : <Navigate to="/login" replace />}</>;
 };
