+++ src/pages/loginPage.js
@@ -1,3 +1,18 @@
+import { useDispatch } from 'react-redux';
+import { toggleLogin } from '../redux/authSlice';
+
 export const LoginPage = () => {
-  return <h3>LoginPage</h3>;
+  const dispatch = useDispatch();
+  const login = () => {
+    dispatch(toggleLogin(true));
+  };
+
+  return (
+    <>
+      <h3>LoginPage</h3>
+      <button type="text" onClick={login}>
+        Login
+      </button>
+    </>
+  );
 };
